Use less API worker with packstack jobs
The number of services is growth in openstack which
leads to higher memory usage. Usually we cannot really leverage
vcpu number of worker threads from each services because multiple
services are active in all time so we just waste time in scheduling.
The nr vpcu strategy would be good if lets say we would run only
keystone on the controller, but in case of packstack jobs
we frequently run compute and volume on the same node as well.

On a 16G memory 8 vcpu machine we also forced to use swap which leads
 to even worse results.

Let's set the default API worker to half of the vcpu number instead of
 the vcpu number, which will lead to less memory usage.

Change-Id: I0d2c8f6b050d2c600103f60284f57e978d2aab9f