Install virtualenv for user only
Often global installation of the virtualenv from pip
can cause conflicts with existing installation of the vitualenv.

So installing virtualenv for user only and use it for pip commands.
Use forceinstall state to ensure that virtualenv will be installed on
user level even when it is already present on global level.

Fixed packstack plugin and added fact collection to get non-root
user home dir and install virtualenv there.

RHOSINFRA-1810

Change-Id: Ic60241ad7aafd8c2a41248b4f7df77c1e4006a8c